requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
cachetools==6.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
cssselect==1.3.0
cssutils==2.11.1
distro==1.9.0
feedparser==6.0.11
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imapclient==3.0.1
jiter==0.10.0
lxml==6.0.0
more-itertools==10.7.0
openai==1.99.3
premailer==3.10.0
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
schedule==1.2.2
sgmllib3k==1.0.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
yagmail==0.15.293
